multiprocessing/requirements.txt
urllib3==1.26.6
python-dotenv
requests

